Move `SnackbarView` to `CommunittyToolkit.Maui.Core`, Require `nfloat` in ctor
