+++ src/CommunityToolkit.Maui/Alerts/Snackbar/Snackbar.macios.cs
@@ -74,62 +74,4 @@ public partial class Snackbar
 	{
 		return new CGRect(cornerRadius.BottomLeft, cornerRadius.TopLeft, cornerRadius.TopRight, cornerRadius.BottomRight);
 	}
-
-	sealed class SnackbarView : ToastView, IDisposable
-	{
-		readonly PaddedButton _actionButton;
-
-		public SnackbarView(
-			string message,
-			UIColor backgroundColor,
-			CGRect cornerRadius,
-			UIColor textColor,
-			UIFont textFont,
-			double characterSpacing,
-			string actionButtonText,
-			UIColor actionTextColor,
-			UIFont actionButtonFont,
-			double padding = DefaultPadding)
-			: base(message, backgroundColor, cornerRadius, textColor, textFont, characterSpacing, padding)
-		{
-			_actionButton = new PaddedButton(padding, padding, padding, padding);
-			ActionButtonText = actionButtonText;
-			ActionTextColor = actionTextColor;
-			ActionButtonFont = actionButtonFont;
-
-			_actionButton.TouchUpInside += ActionButton_TouchUpInside;
-			PopupView.AddChild(_actionButton);
-		}
-
-		public Action? Action { get; init; }
-
-		public string ActionButtonText
-		{
-			get => _actionButton.Title(UIControlState.Normal);
-			private init => _actionButton.SetTitle(value, UIControlState.Normal);
-		}
-
-		public UIColor ActionTextColor
-		{
-			get => _actionButton.TitleColor(UIControlState.Normal);
-			private init => _actionButton.SetTitleColor(value, UIControlState.Normal);
-		}
-
-		public UIFont ActionButtonFont
-		{
-			get => _actionButton.Font;
-			private init => _actionButton.Font = value;
-		}
-
-		void ActionButton_TouchUpInside(object? sender, EventArgs e)
-		{
-			Action?.Invoke();
-			PopupView.Dismiss();
-		}
-
-		public void Dispose()
-		{
-			_actionButton.TouchUpInside -= ActionButton_TouchUpInside;
-		}
-	}
 }
